Correção serviços de Produtos e Criação Alocar Funcionario

+++ src/main/java/com/modulodecompras/modulo/Resources/EstoqueResource.java
@@ -13,7 +13,7 @@ public class EstoqueResource {
     EstoqueService eServ;
 
     @PostMapping("/debitar/{id}/{qtde}")
-    public ResponseEntity<?> debitarProdutoById(@PathVariable int id, @PathVariable int qtde){
+    public ResponseEntity<?> debitarProdutoById(@PathVariable Long id, @PathVariable int qtde){
         try{
             return ResponseEntity.ok(eServ.debitarProdutoPeloId(id,qtde));
         } catch (Exception e) {
@@ -22,11 +22,11 @@ public class EstoqueResource {
     }
 
     @PostMapping("/adicionar/{id}/{qtde}")
-    public ResponseEntity<?> adicionarProdutoById(@PathVariable int id, @PathVariable int qtde) {
+    public ResponseEntity<?> adicionarProdutoById(@PathVariable Long id, @PathVariable int qtde) {
         return ResponseEntity.ok(eServ.adicionarProdutoPeloId(id, qtde));
     }
     @GetMapping("/verificar-epi/{matricula}/{idEPI}")
-    public ResponseEntity<?> verificarEPIPorMatricula(@PathVariable String matricula, @PathVariable int idEPI) {
+    public ResponseEntity<?> verificarEPIPorMatricula(@PathVariable String matricula, @PathVariable Long idEPI) {
         boolean possuiEPI = eServ.verificarEPIPorMatricula(matricula, idEPI);
         if (possuiEPI) {
             return ResponseEntity.ok("O EPI está disponível no estoque.");
